fix: S3 クライアントに proxy 反映済み config を使用
NAT なし VPC 内から S3 へ到達できず画像同期の head/put が失敗していた。
lambda self-invoke と同じ proxy 対応 HTTP クライアントを
infrastructure::aws::load_sdk_config に共通化して両者で使う
